pyproject.toml
[tool.poetry]
name = "personal-finance-bot"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "12.0.0b1"
python-dotenv = "^0.10.1"

[tool.poetry.dev-dependencies]
pylint = "^2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
